clean up html output
