+++ marker/schema/blocks/form.py
@@ -6,4 +6,4 @@ from marker.schema.blocks.basetable import BaseTable
 
 class Form(BaseTable):
     block_type: BlockTypes = BlockTypes.Form
-    block_description: str = "A form, such as a tax form, that contains fields and labels."
+    block_description: str = "A form, such as a tax form, that contains fields and labels.  It most likely doesn't have a table structure."
