+++ marker/schema/blocks/table.py
@@ -4,4 +4,4 @@ from marker.schema.blocks.basetable import BaseTable
 
 class Table(BaseTable):
     block_type: BlockTypes = BlockTypes.Table
-    block_description: str = "A table of data, like a results table."
+    block_description: str = "A table of data, like a results table.  It will be in a tabular format."
