+++ marker/scripts/__init__.py
@@ -1,0 +1,5 @@
+from marker.scripts.convert_single import convert_single_cli
+from marker.scripts.convert import convert_cli
+from marker.scripts.server import server_cli
+from marker.scripts.run_streamlit_app import streamlit_app_cli
+from marker.scripts.chunk_convert import chunk_convert_cli
