+++ marker/scripts/run_streamlit_app.py
@@ -2,12 +2,8 @@ import subprocess
 import os
 
 
-def run():
+def streamlit_app_cli():
     cur_dir = os.path.dirname(os.path.abspath(__file__))
-    app_path = os.path.join(cur_dir, "marker_app.py")
+    app_path = os.path.join(cur_dir, "streamlit_app.py")
     cmd = ["streamlit", "run", app_path]
     subprocess.run(cmd, env={**os.environ, "IN_STREAMLIT": "true"})
-
-
-if __name__ == "__main__":
-    run()
