+++ pyproject.toml
@@ -11,13 +11,7 @@ packages = [
     {include = "marker"}
 ]
 include = [
-    "convert.py",
-    "convert_single.py",
-    "chunk_convert.sh",
-    "chunk_convert.py",
-    "marker_app.py",
-    "run_marker_app.py",
-    "marker_server.py",
+    "marker/scripts/*.sh"
 ]
 
 [tool.poetry.dependencies]
@@ -57,11 +51,11 @@ distance = "0.1.3"
 lxml = "5.3.0"
 
 [tool.poetry.scripts]
-marker = "convert:main"
-marker_single = "convert_single:main"
-marker_chunk_convert = "chunk_convert:main"
-marker_gui = "run_marker_app:run"
-marker_server = "marker_server:main"
+marker = "marker.scripts.convert:convert_cli"
+marker_single = "marker.scripts.convert_single:convert_single_cli"
+marker_chunk_convert = "marker.scripts.chunk_convert:chunk_convert_cli"
+marker_gui = "marker.scripts.run_streamlit_app:streamlit_app_cli"
+marker_server = "marker.scripts.server:server_cli"
 
 [build-system]
 requires = ["poetry-core"]
